+++ src/form/data/mod.rs
@@ -4,7 +4,7 @@ use vertigo::Value;
 use crate::ValidationErrors;
 
 mod data_field;
-pub use data_field::{DataFieldValue, ImageValue, TextAreaValue};
+pub use data_field::{CustomValue, DataFieldValue, ImageValue, TextAreaValue};
 
 mod form_export;
 pub use form_export::{FieldExport, FormExport};